--- FinDataExtractorParser/parser.py
@@ -1,65 +1,0 @@
-import uuid
-from flask import Flask, request, jsonify
-from flask_cors import CORS
-import os
-
-# NOTE
-# comment/uncomment these to save bootup time
-# import llamaTest  
-import gptDataExtraction
-
-app = Flask(__name__)
-CORS(app)
-
-# Temporary folder for uploaded files, possibly the database in the future
-UPLOAD_FOLDER = "uploads"
-os.makedirs(UPLOAD_FOLDER, exist_ok=True)
-
-@app.route('/parse', methods=['POST'])
-def parse_PDF():
-    # Check if the file is part of the request
-    if 'file' not in request.files:
-        return jsonify({"error": "No file part"}), 400
-
-    file = request.files['file']
-
-    if file.filename == '':
-        return jsonify({"error": "No selected file"}), 400
-
-    # generate a unique filename and save file to temperarily (the llama functionality uses a filepath as of now)
-    temp_filename = f"{uuid.uuid4()}.pdf"
-    temp_filepath = os.path.join(UPLOAD_FOLDER, temp_filename)
-    file.save(temp_filepath)
-
-    try:
-        # call main from llamaTest with the file path as the argument
-        # NOTE 
-        # Change the following 2 lines depending on what parser you are needing/testing
-        structured_data = gptDataExtraction.main(temp_filepath)
-            # NOTE THIS DOES COST a small amount of MONEY!!! DO NOT RUN IN A LOOP OR LIKE A MANIAC!!! THANKS! -lukas
-        # structured_data = llamaTest.main(temp_filepath)
-
-        return jsonify({"message": "File uploaded and processed successfully!", "data": structured_data}), 200
-    except Exception as e:
-        return jsonify({"error": str(e)}), 500
-    finally:
-        # delete temporary file
-        if os.path.exists(temp_filepath):
-            os.remove(temp_filepath)
-
-@app.route('/sample', methods=['POST'])
-def print_json():
-    # Get the JSON data from the request
-    data = request.get_json()
-    # Print JSON data to the console
-    print("Received JSON data:", data)
-    # Return a response
-    return jsonify({"status": "success", "message": "Data printed to console"}), 200
-
-@app.route('/sample', methods=['GET'])
-def get_json():
-    # Return a simple message for the GET request
-    return jsonify({"status": "success", "message": "GET request received"}), 200
-
-if __name__ == '__main__':
-    app.run(host='0.0.0.0', port=5000)
