+++ requirements.txt
@@ -4,3 +4,9 @@ Werkzeug==2.0.2
 gunicorn==20.0.0
 pdfplumber
 flask-cors
+# os
+# requests
+# jsonify
+# Ollama
+# uuid
+# llama_cpp
